backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-google-genai
langchain-community
google-search-results
pydantic
python-dotenv
